+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 # [9.8.0] - 2025-01-29
+- Added deprecation notice for Java 8/1.8 users. Version 9.x will be the final version to support Java 8.
 - Messages: Added `trusted_recipient` to SMS, MMS, and MMS
 - Messages: Added support for Typing Indicators in WhatsApp
 - Voice: Websocket Connections can now include custom authorization headers